Class HtmlCleanerApp was created.
#2
