+++ HTML cleanup/HTMLCleanup/HTMLCleanup.csproj
@@ -103,6 +103,7 @@
       <DependentUpon>Config.xsd</DependentUpon>
     </Compile>
     <Compile Include="CleanerConfigSerializer.cs" />
+    <Compile Include="HtmlCleanerApp.cs" />
     <Compile Include="HTMLCleanerInjector.cs" />
     <Compile Include="IntCleanerConfigSerializer.cs" />
     <Compile Include="IntHTMLCleaner.cs" />
